[bat]array.sum(): always create largest type to avoid overflow
Just as with SQL aggregation (aggr.sum()), also SciQL aggregation ([bat]array.sum())
now aggregates into the largest integer (bigint / lng) resp. floating point (double / dbl)
type to avoid overflows (as far as possible).

Extended tests accordingly.

+++ sql/scripts/29_array.sql
@@ -66,8 +66,8 @@ create function array_avg(val real,     dim1 int, offsets1 int, size1 int) retur
 create function array_avg(val double,   dim1 int, offsets1 int, size1 int) returns double external name "array".avg;
 
 create function array_sum(val tinyint,  dim1 int, offsets1 int, size1 int) returns bigint external name "array".sum;
-create function array_sum(val smallint, dim1 int, offsets1 int, size1 int) returns double external name "array".sum;
-create function array_sum(val int,      dim1 int, offsets1 int, size1 int) returns double external name "array".sum;
+create function array_sum(val smallint, dim1 int, offsets1 int, size1 int) returns bigint external name "array".sum;
+create function array_sum(val int,      dim1 int, offsets1 int, size1 int) returns bigint external name "array".sum;
 create function array_sum(val bigint,   dim1 int, offsets1 int, size1 int) returns bigint external name "array".sum;
 create function array_sum(val real,     dim1 int, offsets1 int, size1 int) returns double external name "array".sum;
 create function array_sum(val double,   dim1 int, offsets1 int, size1 int) returns double external name "array".sum;
@@ -80,8 +80,8 @@ create function array_avg(val real,     dim1 int, offsets1 int, size1 int, dim2 
 create function array_avg(val double,   dim1 int, offsets1 int, size1 int, dim2 int, offsets2 int, size2 int) returns double external name "array".avg;
 
 create function array_sum(val tinyint,  dim1 int, offsets1 int, size1 int, dim2 int, offsets2 int, size2 int) returns bigint external name "array".sum;
-create function array_sum(val smallint, dim1 int, offsets1 int, size1 int, dim2 int, offsets2 int, size2 int) returns double external name "array".sum;
-create function array_sum(val int,      dim1 int, offsets1 int, size1 int, dim2 int, offsets2 int, size2 int) returns double external name "array".sum;
+create function array_sum(val smallint, dim1 int, offsets1 int, size1 int, dim2 int, offsets2 int, size2 int) returns bigint external name "array".sum;
+create function array_sum(val int,      dim1 int, offsets1 int, size1 int, dim2 int, offsets2 int, size2 int) returns bigint external name "array".sum;
 create function array_sum(val bigint,   dim1 int, offsets1 int, size1 int, dim2 int, offsets2 int, size2 int) returns bigint external name "array".sum;
 create function array_sum(val real,     dim1 int, offsets1 int, size1 int, dim2 int, offsets2 int, size2 int) returns double external name "array".sum;
 create function array_sum(val double,   dim1 int, offsets1 int, size1 int, dim2 int, offsets2 int, size2 int) returns double external name "array".sum;
